Tests.txt corpus sweep (2170 positions) + puzzle suite to 9 - FenCorpus.TestsTxtInvariants: every FEN ever noted must import (fen_ok), generate moves and evaluate sanely, with color-mirror symmetry checked over real middlegames (tolerance +-2cp: residual float summation-order noise flips an int truncation by 1cp; true indexing bugs fail hard). get_long_term_piece_mobility accumulators float->double kills most of that noise at the source. Ladders now 9 puzzles (Ne7+ royal fork REBALANCED to B+N vs N - first version was Q+R vs N so the fork lost anyway and the engine was RIGHT to refuse it; Nf7# smothered). Baselines: 9/9 at 3s, 600 nodes total (fork back at the 25-node floor). Also: unit_test.cpp EOL normalization (2501 mixed CRLF were shifting compiler line maps) and two mojibake-blasted comment lines repaired

+++ opti_chess/eval_mobility.cpp
@@ -1705,7 +1705,7 @@ int Board::get_long_term_piece_mobility(bool display) const {
 		const uint8_t col = sq & 7;
 		uint8_t piece = _array[row][col];
 
-		float piece_mobility = 0.0f;
+		double piece_mobility = 0.0;
 
 		// White pawn
 		if (piece == w_pawn) {
@@ -1737,7 +1737,7 @@ int Board::get_long_term_piece_mobility(bool display) const {
 					int new_row = row + d_row;
 					int new_col = col + d_col;
 
-					float cumulative_blocking_factor = 1.0f;
+					double cumulative_blocking_factor = 1.0;
 
 					while (is_in(new_row, 0, 7) && is_in(new_col, 0, 7)) {
 						uint8_t p = piece_at(new_row, new_col);
@@ -1769,7 +1769,7 @@ int Board::get_long_term_piece_mobility(bool display) const {
 					int new_row = row + d_row;
 					int new_col = col + d_col;
 
-					float cumulative_blocking_factor = 1.0f;
+					double cumulative_blocking_factor = 1.0;
 
 					while (is_in(new_row, 0, 7) && is_in(new_col, 0, 7)) {
 						uint8_t p = piece_at(new_row, new_col);
@@ -1835,7 +1835,7 @@ int Board::get_long_term_piece_mobility(bool display) const {
 					int new_row = row + d_row;
 					int new_col = col + d_col;
 
-					float cumulative_blocking_factor = 1.0f;
+					double cumulative_blocking_factor = 1.0;
 
 					while (is_in(new_row, 0, 7) && is_in(new_col, 0, 7)) {
 						uint8_t p = piece_at(new_row, new_col);
@@ -1867,7 +1867,7 @@ int Board::get_long_term_piece_mobility(bool display) const {
 					int new_row = row + d_row;
 					int new_col = col + d_col;
 
-					float cumulative_blocking_factor = 1.0f;
+					double cumulative_blocking_factor = 1.0;
 
 					while (is_in(new_row, 0, 7) && is_in(new_col, 0, 7)) {
 						uint8_t p = piece_at(new_row, new_col);
